requirements.txt
click==7.1.2
numpy==1.20.1
python-dotenv==0.21.0
setuptools==52.0.0
torch==1.13.1
torchmetrics==0.11.0
